made disk implementations bulletproof

+++ ljson/slapdash/disk.py
@@ -8,12 +8,14 @@ import json, os
 
 class Table(SlapdashTable):
 	def __init__(self, header, file_, reread_stats = False):
-		self._first_next_call = True
 		self.header = header
 		self.file = file_
 
 		self._index = 0
 		self.document_count = 0
+
+		self._file_pointer_stack = deque()
+		self._loop_file_pointer_stack = deque()
 
 		if(reread_stats):
 			self.insert_stats()
@@ -25,16 +27,30 @@ class Table(SlapdashTable):
 		except KeyError:
 			pass
 
-	@staticmethod
-	def open(filename):
+	
+
+
+	@classmethod
+	def empty(cls, file_or_filename):
 		"""
-		Equivalent to ``Table.from_file(open(filename, "r+"))``
+		Create a new empty table with the given file
+		or filename.
 		"""
-		if(not os.path.exists(filename)):
-			raise IOError("cannot open {} for reading: does not exist".format(filename))
-		fin = open(filename, "r+")
-		header = SlapdashHeader.from_file(fin)
-		return Table(header, fin)
+
+		header = SlapdashHeader({})
+		
+		if(isinstance(file_or_filename, str)):
+			file_ = open(file_or_filename, "w+")
+		else:
+			file_ = file_or_filename
+
+		file_.write(header.get_header())
+		table = cls(header, file_)
+
+		table.insert_stats()
+
+		return table
+
 
 
 	@staticmethod
@@ -46,7 +62,7 @@ class Table(SlapdashTable):
 		return Table(header, file_)
 
 	def additem(self, document):
-		self._first_next_call = True
+		seek_after = self.file.tell()
 		if(not "field_count" in self.header.descriptor):
 			self.insert_stats()
 		old_data = {k: self.header.descriptor[k] 
@@ -75,6 +91,7 @@ class Table(SlapdashTable):
 				wrong_type = (k, v)
 				
 		if(raise_key_error):
+			self.file.seek(seek_after)
 			raise KeyError("Unknown datatype for field {}: {}".format(*wrong_type))
 
 		self.document_count += 1
@@ -86,40 +103,59 @@ class Table(SlapdashTable):
 		self.file.seek(0, 2) # EOF here
 		self.file.write("\n")
 		json.dump(document, self.file)
+		self.file.seek(seek_after)
 
 	def __getitem__(self, dct):
 		return Selector(self.header, dct, self)
 	def __iter__(self):
-		self._first_next_call = True
+		self._file_pointer_stack.appendleft(self.file.tell())
+		self.file.seek(0)
+		self.file.readline()
+		self._loop_file_pointer_stack.appendleft(self.file.tell())
 		return self
 	def __contains__(self, dct):
-		self._first_next_call = True
+		seek_after = self.file.tell()
 		self.file.seek(0)
 		self.file.readline()
 		for row in self.file:
 			if(row.isspace()):
 				continue
 			if(document_matches(json.loads(row), dct)):
+				self.file.seek(seek_after)
 				return True
+		self.file.seek(seek_after)
 		return False
 	def save(self, fout):
 		"""
 		Save this table to the file ``fout``.
 		"""
-		self._first_next_call = True
+		seek_after = self.file.tell()
 		self.file.seek(0)
 		for r in self.file:
 			if(r.strip == ""):
 				continue
 			fout.write(r)
+		self.file.seek(seek_after)
 	def __next__(self):
-		if(self._first_next_call):
-			self._first_next_call = False
-			self.file.seek(0)
-			self.file.readline()
-		row = self.file.__next__()
-		while(row.strip() == ""):
-			row = self.file.__next__()
+		seek_after = self.file.tell()
+		self.file.seek(self._loop_file_pointer_stack.popleft())
+
+
+		row = self.file.readline()
+		if(not row):
+			# we are leaving the loop => this context is done
+			# => return to the context before __iter__ was called
+			self.file.seek(self._file_pointer_stack.popleft())
+			raise StopIteration()
+		while(row.isspace()):
+			if(not row):
+				# we are leaving the loop => this context is done
+				# => return to the context before __iter__ was called
+				self.file.seek(self._file_pointer_stack.popleft())
+				raise StopIteration()
+			row = self.file.readline()
+		self._loop_file_pointer_stack.appendleft(self.file.tell())
+		self.file.seek(seek_after)
 		return json.loads(row)
 
 	def __enter__(self):
@@ -132,6 +168,7 @@ class Table(SlapdashTable):
 
 
 	def insert_stats(self):
+		seek_after = self.file.tell()
 		self.header.descriptor.update(self.calculate_stats())
 
 		os.unlink(self.file.name)
@@ -147,8 +184,11 @@ class Table(SlapdashTable):
 			new_file.write(line)
 		self.file.close()
 		self.file = new_file
+		self.file.seek(seek_after)
+
 
 	def calculate_stats(self):
+		seek_after = self.file.tell()
 		counter = defaultdict(int)
 		dtype_counter = defaultdict(int)
 		dtype_per_field_counter = defaultdict()
@@ -161,11 +201,16 @@ class Table(SlapdashTable):
 					dtype_per_field_counter[k] = defaultdict(int)
 				dtype_per_field_counter[k][inversed_datatypes[type(v)]] += 1
 
+		self.file.seek(seek_after)
 		return {"length": self.document_count, "field_count": dict(counter),
 			"total_datatype_count": dict(dtype_counter),
 			"per_field_datatype_count": dict(dtype_per_field_counter)}
 	def __delitem__(self, dct):
-		self._first_next_call = True
+		if(self._file_pointer_stack):
+			raise RuntimeError("Attempt to delete item from table during looping. This leads to undefined behaviour.")
+		seek_after = self.file.tell()
+		max_seek = 0
+
 		self.file.seek(0)
 		os.unlink(self.file.name)
 		buf = open(self.file.name, "w+")
@@ -180,9 +225,15 @@ class Table(SlapdashTable):
 				self.document_count -= 1
 			else:
 				buf.write(line)
+				max_seek = buf.tell()
 		buf.seek(0)
 		self.file.close()
 		self.file = buf
+		if(seek_after <= max_seek):
+			self.file.seek(seek_after)
+		else:
+			self.file.seek(0)
+			raise RuntimeWarning("File is truncated. unable to restore table state.")
 		if(not deleted_row):
 			raise KeyError("no matching rows found: {}".format(dct))
 
@@ -193,24 +244,26 @@ class Selector(object):
 		self.file = table.file
 		self.dct = dct
 		self.table = table
-		self._first_next_call = True
+		self._file_pointer_stack = deque()
+		self._loop_file_pointer_stack = deque()
 
 	def getone(self, column = None):
-		self._first_next_call = True
-		self.table._first_next_call = True
+		seek_after = self.file.tell()
 		self.file.seek(0)
 		self.file.readline()
 		for line in self.file:
 			row = json.loads(line)
 			if(document_matches(row, self.dct)):
 				if(column):
+					self.file.seek(seek_after)
 					return row[column]
 				else:
+					self.file.seek(seek_after)
 					return row
+		self.file.seek(seek_after)
 		return None
 	def __getitem__(self, column):
-		self._first_next_call = True
-		self.table._first_next_call = True
+		seek_after = self.file.tell()
 		self.file.seek(0)
 		self.file.readline()
 		res = deque()
@@ -218,11 +271,13 @@ class Selector(object):
 			row = json.loads(line)
 			if(document_matches(row, self.dct)):
 				res.append(row[column])
+		self.file.seek(seek_after)
 		return list(res)
 			
 	def __setitem__(self, column, value):
-		self._first_next_call = True
-		self.table._first_next_call = True
+		if(self._file_pointer_stack or self.table._file_pointer_stack):
+			raise RuntimeError("Attempt to edit table while iterating")
+		seek_after = self.file.tell()
 		self.file.seek(0)
 		os.unlink(self.file.name)
 		buf = open(self.file.name, "w+")
@@ -240,22 +295,36 @@ class Selector(object):
 		self.file.close()
 		self.file = buf
 		self.table.file = buf
+		self.table.file.seek(seek_after)
 
 	def __next__(self):
-		self.table._first_next_call = True
-		if(self._first_next_call):
-			self._first_next_call = False
-			self.file.seek(0)
-			self.file.readline()
-		row = self.file.__next__()
+		seek_after = self.file.tell()
+		self.file.seek(self._loop_file_pointer_stack.popleft())
+
+
+		row = self.file.readline()
+		if(not row):
+			# we are leaving the loop => this context is done
+			# => return to the context before __iter__ was called
+			self.file.seek(self._file_pointer_stack.popleft())
+			raise StopIteration()
 		data = {}
 		if(not row.isspace()):
 			data = json.loads(row)
 		while(row.isspace() or not document_matches(data, self.dct)):
-			row = self.file.__next__()
+			row = self.file.readline()
+			if(not row):
+				# we are leaving the loop => this context is done
+				# => return to the context before __iter__ was called
+				self.file.seek(self._file_pointer_stack.popleft())
+				raise StopIteration()
 			data = json.loads(row) if not row.isspace() else {}
+		self._loop_file_pointer_stack.appendleft(self.file.tell())
+		self.table.file.seek(seek_after)
 		return data
 	def __iter__(self):
-		self.table._first_next_call = True
-		self._first_next_call = True
+		self._file_pointer_stack.appendleft(self.file.tell())
+		self.file.seek(0)
+		self.file.readline()
+		self._loop_file_pointer_stack.appendleft(self.file.tell())
 		return self
